lamp_admin_mcp/pyproject.toml
[project]
name = "lamp_admin_cli"
classifiers = ["Private :: Do Not Upload"]
dynamic = ["version"]
authors = [
    {name="Jeuk Oh", email="juoh@letsur.ai"},
    {name="Jaehyun Song", email="jhsong@letsur.ai"},
]
description = "Python package for easy uploading model to lamp by ml engineer in Letsur"
readme = "README.md"

requires-python = ">=3.10"
dependencies = [
    "boto3>=1.35.99",
    "botocore>=1.35.99",
    "certifi==2023.11.17",
    "charset-normalizer==3.3.2",
    "cloudpickle==3.0.0",
    "databricks-cli==0.18.0",
    "docker==7.1.0",
    "entrypoints==0.4",
    "fastapi>=0.104.0",
    "gitdb==4.0.11",
    "gitpython==3.1.41",
    "idna>=3.6",
    "importlib-metadata==7.0.1",
    "jinja2==3.1.3",
    "jmespath==1.0.1",
    "markdown-it-py==3.0.0",
    "mcp[cli]>=1.17.0",
    "mdurl==0.1.2",
    "pydantic>=2.7.0",
    "mlflow-skinny==2.16.0",
    "oauthlib==3.2.2",
    "packaging==23.2",
    "pandas-stubs==2.3.2.250926",
    "protobuf==4.25.2",
    "pygithub==2.3.0",
    "pygments==2.17.2",
    "pyjwt==2.8.0",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.1",
    "python-multipart>=0.0.6",
    "pytz==2023.3.post1",
    "pyyaml==6.0.1",
    "requests==2.31.0",
    "rich==13.7.0",
    "s3transfer==0.10.0",
    "six==1.16.0",
    "smmap==5.0.1",
    "sqlparse==0.4.4",
    "tabulate==0.9.0",
    "urllib3>=1.26.20",
    "uvicorn>=0.24.0",
    "zipp==3.17.0",
]

[project.optional-dependencies]
dev = [
    "boto3-stubs>=1.38.14",
    "mypy-boto3-s3>=1.38.0",
    "pytest==7.4.4",
]

[project.scripts]
# MCP-only branch: no CLI entrypoints

[project.urls]
repository = "https://github.com/letsur-dev/lamp_admin_cli"


[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]


[tool.uv]
default-groups = ["dev"]

[dependency-groups]
dev = [
    "boto3-stubs>=1.38.14",
    "mypy-boto3-s3>=1.38.0",
    "setuptools-scm==8.3.1",
    "pytest==7.4.4",
]
